+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- *Breaking*: remove possibility to GET `/api/entries/:id`, just use `/:id`.
 - Return appropriate content type for `/:id` based on `accept` header (i.e.
   `text/html` returns the HTML page) and the `fmt` query parameter (i.e. set to
   `raw` returns raw text).